+++ index.html
@@ -7,7 +7,8 @@
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
 
-  <script defer src="js/main.js"></script>
+<!--  <script defer src="js/main.js"></script>-->
+  <script type="module" src="js/main.js"></script>
 
   <title>Кексобукинг</title>
 </head>
