+++ js/utils/numbers-generation.js
@@ -1,0 +1,29 @@
+
+// Функция взята из интернета и доработана
+// Источник - https://github.com/you-dont-need/You-Dont-Need-Lodash-Underscore#_random
+function getRandomPositiveFloat(arg1, arg2, digits = 1) {
+  // Чтобы не заставлять пользователя нашей функции помнить порядок аргументов,
+  // реализуем поддержку передачи минимального и максимального значения в любом порядке,
+  // а какое из них большее и меньшее вычислим с помощью Math.min и Math.max
+  const lower = Math.min(Math.abs(arg1), Math.abs(arg2));
+  const upper = Math.max(Math.abs(arg1), Math.abs(arg2));
+  const result = Math.random() * (upper - lower) + lower;
+  return result.toFixed(digits);
+}
+
+function getRandomPositiveInteger (arg1, arg2) {
+  // Чтобы не заставлять пользователя нашей функции помнить порядок аргументов,
+  // реализуем поддержку передачи минимального и максимального значения в любом порядке,
+  // а какое из них большее и меньшее вычислим с помощью Math.min и Math.max.
+
+  // После нам нужно убедиться, что пользователь не передал дробные значения,
+  // для этого на всякий пожарный случай нижнюю границу диапазона
+  // мы округляем к ближайшему большему целому с помощью Math.ceil,
+  // а верхнюю границу - к ближайшему меньшему целому с помощью Math.floor
+  const lower = Math.ceil(Math.min(Math.abs(arg1), Math.abs(arg2)));
+  const upper = Math.floor(Math.max(Math.abs(arg1), Math.abs(arg2)));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+}
+
+export {getRandomPositiveFloat, getRandomPositiveInteger};
